Fix int8 overflow error with large hypervector dimensions
The issue was that dimension values like 8192 were being used with np.int8
arrays, which can only hold values from -128 to 127. When 8192 was converted
to int8, it would overflow to 0, causing vectors to have shape (0,) instead
of (8192,).

Changes:
- Changed from int8 to int16 for bipolar hypervectors in hv_space.py
- Updated performance.py to use int16 for bipolar vectors
- Adjusted memory estimation to account for int16 (2 bytes per element)

This fixes the 'Python integer 8192 out of bounds for int8' error reported
during CI/CD runs.

+++ tailchasing/performance.py
@@ -41,7 +41,8 @@ class OptimizedHVSpace:
     def _rand_vec(self) -> np.ndarray:
         """Generate random vector."""
         if self.bipolar:
-            return np.random.choice([-1, 1], size=self.dim).astype(np.int8)
+            # Use int16 to avoid overflow issues with large dimensions
+            return np.random.choice([-1, 1], size=self.dim).astype(np.int16)
         else:
             return np.random.randint(0, 2, size=self.dim, dtype=np.uint8)
     
@@ -95,9 +96,9 @@ class OptimizedHVSpace:
         
         # For bipolar, use sign of sum
         else:
-            stacked = np.array(vectors, dtype=np.int8)
+            stacked = np.array(vectors, dtype=np.int16)
             sums = np.sum(stacked, axis=0)
-            return np.sign(sums).astype(np.int8)
+            return np.sign(sums).astype(np.int16)
 
 
 class ParallelSemanticAnalyzer:
